Update get_traject.m
I suspect there was a mistake in the formula I put in the lecture. The equation for f (although it be simplified further) is the correct one.

Something weird that remain is that the reconstructed values go well beyond 0 and 1 (I used ylim([0,1]) to plot the result). It does not worry me too much as I encountered similar issues with my method.

About my method, I am not surprised that it gives different results as it is based on a previous version of DEB (pre-2010). However, I wonder if not having daily values for growth rates and length makes it less accurate. I just presented you a simplified version of this reconstruction method too.
I would use the first approach in this case, just to be sure.

Romain